+++ config/dev.exs
@@ -3,14 +3,11 @@ import Config
 # Clear the terminal when running mix test.watch
 config :mix_test_watch, clear: true
 
-# Define default any_http client adapter
-config :any_http, client_adapter: AnyHttp.Adapters.Httpc
-
 # Configure ElasticsearchEx
 config :elasticsearch_ex,
   clusters: %{
     default: %{
       endpoint: "https://elastic:elastic@localhost:9200",
-      http_opts: [ssl: [verify: :verify_none]]
+      req_opts: [connect_options: [transport_opts: [verify: :verify_none]]]
     }
   }